# xt_server_openApi_autogenerated.json
{
  "swagger": "2.0",
  "info": {
    "title": "XTVision server",
    "description": "All XTVsion server apis",
    "version": "1.0.0"
  },
  "host": "localhost:5001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "paths": {
...
    "/roles/": {
      "post": {
        "tags": [
          "ROLES"
        ],
        "summary": "add a role",
        "description": "",
        "parameters": [
          {
            "name": "NewRole",
            "in": "body",
            "description": "role to add",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "featureIds": {
                  "type": "array",
                  "example": [
                    1
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "userIds": {
                  "type": "array",
                  "example": [
                    1
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Returned role",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 2
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "featureIds": {
                  "type": "array",
                  "example": [
                    1
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "userIds": {
                  "type": "array",
                  "example": [
                    1
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "id",
                "name"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ROLES"
        ],
        "summary": "get all roles",
        "description": "",
        "responses": {
          "200": {
            "description": "Returned roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 2
                  },
                  "name": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "string"
                  },
                  "featureIds": {
                    "type": "array",
                    "example": [
                      1
                    ],
                    "items": {
                      "type": "number"
                    }
                  },
                  "userIds": {
                    "type": "array",
                    "example": [
                      1
                    ],
                    "items": {
                      "type": "number"
                    }
                  }
                },
                "required": [
                  "id",
                  "name"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ROLES"
        ],
        "summary": "delete all roles or by name",
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "type": "string",
            "description": "role name."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/roles/{id}": {
      "put": {
        "tags": [
          "ROLES"
        ],
        "summary": "insert or update a role",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "role ID."
          },
          {
            "name": "user",
            "in": "body",
            "description": "user to add or to update",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "featureId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "ROLES"
        ],
        "summary": "delete one role",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "role ID."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/roles/linkUsers": {
      "post": {
        "tags": [
          "ROLES"
        ],
        "summary": "link a role to n users",
        "description": "",
        "parameters": [
          {
            "name": "NewRoleUser",
            "in": "body",
            "description": "role id and users id",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "RoleId": {
                    "type": "number",
                    "example": 1
                  },
                  "UsersId": {
                    "type": "number",
                    "example": 1
                  }
                },
                "required": [
                  "RoleId",
                  "UsersId"
                ]
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/roles/linkFeatures": {
      "post": {
        "tags": [
          "ROLES"
        ],
        "summary": "link role to features",
        "description": "",
        "parameters": [
          {
            "name": "NewRoleFeature",
            "in": "body",
            "description": "role id and features id",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "RoleId": {
                    "type": "number",
                    "example": 1
                  },
                  "FeaturesId": {
                    "type": "number",
                    "example": 1
                  }
                },
                "required": [
                  "RoleId",
                  "FeaturesId"
                ]
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/databases/{backupNames}": {
      "delete": {
        "tags": [
          "DATABASE"
        ],
        "summary": "Delete a manual backup",
        "description": "",
        "parameters": [
          {
            "name": "backupNames",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "backup name"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}